requirements.txt
python-telegram-bot==13.15
requests==2.32.5
python-dotenv==1.2.1
Flask==2.3.4
